seperated RescuWise into seperate apps to assit beginner developers
