+++ RescuWise/urls.py
@@ -14,44 +14,23 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 
-from django.contrib import admin
-from django.urls import path
-from RescuWise.views import *
-from django.contrib.auth.decorators import login_required
+
+from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
 
-from django.contrib.auth import views as auth_views
-from django.contrib.auth.decorators import login_required
-from django.contrib.staticfiles.urls import staticfiles_urlpatterns
-
 urlpatterns = [
-    # temp as example
-    path('animals/', AnimalIndex.as_view(), name='animals'),
-
-    # Admin Urls
-    path('admin/', admin.site.urls),
     
-    # Login and Auth Views #
-    path('signup/', SignUp.as_view(), name="signup"),
-    path('login/', Login.as_view(), name="login"),
-    path('logout/', Logout.as_view(), name="logout"),
-    path('password_reset/', auth_views.PasswordResetView.as_view(template_name='password_reset.html'), name='password_reset'),
-    path('password_reset/<uidb64>/<token>/', auth_views.PasswordResetConfirmView.as_view(template_name='password_reset_confirm.html'), name='password_reset_confirm'),
-    path('password_reset/done/', auth_views.PasswordResetDoneView.as_view(template_name='password_reset_done.html'), name='password_reset_done'),
-    path('password_reset/complete/', auth_views.PasswordResetCompleteView.as_view(template_name='password_reset_complete.html'), name='password_reset_complete'),
-    path('access_resource_error/', TemplateView.as_view(template_name="access_resource_error.html")),
+    # include core app urls (static/non-auth pages)
+    path('', include('core.urls')),
 
-    path('dashboard/', DashboardView.as_view(), name='dashboard'),
-    
-    # action urls
-    path('createanimal/',CreateAnimal.as_view()),
-    path('listanimals/',ListAnimals.as_view()),
-    path('createshelter/',CreateShelter.as_view()),
-    path('listshelters/',ListShelters.as_view()),
+    # include shelters app urls
+    path('', include('shelters.urls')),
 
-    # catch alls
-    path('', Home.as_view()),
-    path('<template_name>/', DynamicStaticPages.as_view()),
-    
+    # include animals app urls
+    path('', include('animals.urls')),
+
+    # include authentication urls
+    path('', include('authentication.urls')),
+
 ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
