+++ core/admin.py
@@ -1,6 +1,10 @@
 from django.contrib import admin
-from .models import *
+from animals.models import *
+from shelters.models import *
+from authentication.models import *
+from core.models import *
 
 # Register your models here.
 admin.site.register(Animal)
 admin.site.register(Shelter)
+admin.site.register(Profile)
